requirements.txt
asgiref==3.5.0
Django==4.0.1
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.1.0
